restructure initialize
