+++ src/salamander/nmf_framework/mvnmf.py
@@ -5,15 +5,15 @@ from typing import TYPE_CHECKING
 import numpy as np
 from numba import njit
 
+from ..initialization.initialize import EPSILON
 from ..utils import normalize_WH
 from ._utils_klnmf import kl_divergence, samplewise_kl_divergence, update_H
-from .initialization import EPSILON
 from .standard_nmf import StandardNMF
 
 if TYPE_CHECKING:
     from typing import Any, Literal
 
-    from .initialization import _Init_methods
+    from ..initialization.methods import _Init_methods
 
 _DEFAULT_FITTING_KWARGS = None
